feat(optional-object-attribute.mdx): add new TIL about using the spread operator for conditional attributes in JavaScript
This commit adds a new TIL (Today I Learned) file that explains how to add optional attributes to an object conditionally in JavaScript. The traditional approach of manipulating the object after its declaration is compared to a more elegant solution using the spread operator. The spread operator allows for the condition to be incorporated directly into the object declaration, resulting in cleaner and more readable code. The commit also includes an example and a link to a Twitter post where this solution was found.